Introduce new classes for task and tasklist management.
This also has a dependency on Python 3.7 since there are type
annotations that require forward declarations.

These new classes are not yet used by the main Tasky module.

This commit also moves the TextColor class into this module to avoid
future circular dependencies between the main module and this one.