Use the new classes for task and tasklist management.
This commit also enforces a new restriction against tasks greater
than one level of nesting. This is consistent with a recent change in
the actual Google Tasks product which will also be enforced in the API
in the future.

This rewrite is still in experimental shape so there may be bugs.